requirements.txt
Django==3.0.7
djangorestframework==3.11.0
django-oauth-toolkit==1.3.2